Add local bfd test
This PR  adds test for the BFD feature of FRR,
The test set-up configamps with the needed FRR configurations and
deploys with daemonset FRR containers on each node on the cluster,
peering it with the "external router" which is a master node.
The test asserts that the status of all the bfd sessions is up.