Add support for HTTP transport type
Hw-event-proxy supports HTTP transport type starting
at 4.12. AMQP transport type is still supported for
now but will be obsolete in the future. The transport
types have to be consistent at sending and receiving sides
for the event framework to work.

With this change the automation script queries
the transport type from hw-event-proxy deployment and
deploy consumer with the same transport type.

Signed-off-by: Jack Ding <jackding@gmail.com>