Print consumer log after each test
Consumer is launched in the test suite setup, but the consumer
test case only ran towards the end of the test run in CI. We
need more logs to identify when consumer stopped receiving events.
The cause of the consumer test failure could be any of the
test case executed prior to the failed test.